deeplearning_team/Meng/requirements.txt
# Install PyTorch with CUDA 12.6 support first (not in this file):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126
torch>=2.1.0
torchvision>=0.16.0

# Data loading
kagglehub>=1.0.0
kaggle>=2.1.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0

# Image / GIF output
pillow>=10.0.0
imageio>=2.37.0
imageio-ffmpeg>=0.6.0

# Training utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.14.0
matplotlib>=3.7.0

# Sentence-level translation
transformers>=4.40.0
sacrebleu>=2.3.0
